+++ src/handlers/mod.rs
@@ -1,0 +1,21 @@
+pub mod erase_handler;
+pub mod install_handler;
+pub mod list_handler;
+pub mod rollback_handler;
+pub mod sync_handler;
+pub mod uninstall_handler;
+pub mod use_handler;
+
+use super::version::types::LocalVersion;
+
+
+pub enum InstallResult {
+    InstallationSuccess(String),
+    VersionAlreadyInstalled,
+    NightlyIsUpdated,
+}
+
+pub enum PostDownloadVersionType {
+    Standard(LocalVersion),
+    Hash,
+}
