+++ src/handlers/rollback_handler.rs
@@ -1,11 +1,14 @@
-use super::{use_handler, utils};
-use crate::models::{Config, LocalNightly, Nightly};
+use crate::helpers::version::nightly::produce_nightly_vec;
 use anyhow::Result;
 use chrono::{Duration, Utc};
 use dialoguer::{console::Term, theme::ColorfulTheme, Select};
-use regex::Regex;
-use tokio::fs;
 use tracing::info;
+
+use crate::{
+    config::Config,
+    handlers::use_handler,
+    helpers::{self, version::types::ParsedVersion},
+};
 
 pub async fn start(config: Config) -> Result<()> {
     let nightly_vec = produce_nightly_vec(&config).await?;
@@ -32,7 +35,7 @@ pub async fn start(config: Config) -> Result<()> {
 
     match selection {
         Some(i) => {
-            let is_version_used = utils::is_version_used(&name_list[i], &config).await;
+            let is_version_used = helpers::version::is_version_used(&name_list[i], &config).await;
 
             if is_version_used {
                 info!("{} is already used.", &name_list[i]);
@@ -41,9 +44,9 @@ pub async fn start(config: Config) -> Result<()> {
 
             use_handler::switch(
                 &config,
-                &crate::models::InputVersion {
+                &ParsedVersion {
                     tag_name: name_list[i].clone(),
-                    version_type: crate::enums::VersionType::Standard,
+                    version_type: crate::helpers::version::types::VersionType::Standard,
                 },
             )
             .await?;
@@ -73,41 +76,6 @@ pub async fn start(config: Config) -> Result<()> {
     }
 
     Ok(())
-}
-
-pub async fn produce_nightly_vec(config: &Config) -> Result<Vec<LocalNightly>> {
-    let downloads_dir = utils::get_downloads_folder(config).await?;
-    let mut paths = fs::read_dir(&downloads_dir).await?;
-
-    let regex = Regex::new(r"nightly-[a-zA-Z0-9]{8}")?;
-
-    let mut nightly_vec: Vec<LocalNightly> = Vec::new();
-
-    while let Some(path) = paths.next_entry().await? {
-        let name = path.file_name().into_string().unwrap();
-
-        if !regex.is_match(&name) {
-            continue;
-        }
-
-        let nightly_content = path.path().join("bob.json");
-        let nightly_string = fs::read_to_string(nightly_content).await?;
-
-        let nightly_data: Nightly = serde_json::from_str(&nightly_string)?;
-
-        let mut nightly_entry = LocalNightly {
-            data: nightly_data,
-            path: path.path(),
-        };
-
-        nightly_entry.data.tag_name = name;
-
-        nightly_vec.push(nightly_entry);
-    }
-
-    nightly_vec.sort_by(|a, b| b.data.published_at.cmp(&a.data.published_at));
-
-    Ok(nightly_vec)
 }
 
 fn humanize_duration(duration: Duration) -> Result<String> {
