+++ src/handlers/uninstall_handler.rs
@@ -1,19 +1,20 @@
-use crate::{models::Config, modules::utils};
 use anyhow::{anyhow, Result};
 use reqwest::Client;
 use tokio::fs;
 use tracing::{info, warn};
 
+use crate::{config::Config, helpers::{self, directories}};
+
 pub async fn start(version: &str, config: Config) -> Result<()> {
     let client = Client::new();
-    let version = utils::parse_version_type(&client, version).await?;
+    let version = helpers::version::parse_version_type(&client, version).await?;
 
-    if utils::is_version_used(&version.tag_name, &config).await {
+    if helpers::version::is_version_used(&version.tag_name, &config).await {
         warn!("Switch to a different version before proceeding");
         return Ok(());
     }
 
-    let downloads_dir = match utils::get_downloads_folder(&config).await {
+    let downloads_dir = match directories::get_downloads_directory(&config).await {
         Ok(value) => value,
         Err(error) => return Err(anyhow!(error)),
     };
