+++ src/main.rs
@@ -1,18 +1,19 @@
-mod enums;
-mod models;
-mod modules;
+mod cli;
+mod config;
+mod handlers;
+mod helpers;
 
 extern crate core;
 
 use anyhow::{anyhow, Result};
-use models::Config;
-use modules::utils;
-use regex::Regex;
+use config::{handle_config, Config};
+use helpers::version;
 use std::{
     env,
     process::{exit, Command}, path::Path,
 };
 use tracing::{error, Level};
+use crate::helpers::directories;
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -29,8 +30,7 @@ async fn main() -> Result<()> {
 }
 
 async fn run() -> Result<()> {
-    let config_file = crate::modules::fs::get_config_dir()?;
-    let config: Config = handle_config(tokio::fs::read_to_string(config_file).await)?;
+    let config: Config = handle_config().await?;
 
     let args: Vec<String> = env::args().collect();
 
@@ -41,9 +41,9 @@ async fn run() -> Result<()> {
     if !exe_name.contains("bob") {
         let rest_args = &args[1..];
 
-        let downloads_dir = utils::get_downloads_folder(&config).await?;
-        let platform = utils::get_platform_name();
-        let used_version = utils::get_current_version(&config).await?;
+        let downloads_dir = directories::get_downloads_directory(&config).await?;
+        let platform = helpers::get_platform_name();
+        let used_version = version::get_current_version(&config).await?;
 
         let location = downloads_dir
             .join(used_version)
@@ -68,55 +68,6 @@ async fn run() -> Result<()> {
         }
     }
 
-    modules::cli::start(config).await?;
-    Ok(())
-}
-
-fn handle_config(config_file: Result<String, std::io::Error>) -> Result<Config> {
-    let config = match config_file {
-        Ok(config_file) => {
-            let mut config: Config = serde_json::from_str(&config_file)?;
-            handle_envars(&mut config)?;
-            config
-        }
-        Err(_) => Config {
-            enable_nightly_info: None,
-            downloads_location: None,
-            installation_location: None,
-            version_sync_file_location: None,
-            rollback_limit: None,
-        },
-    };
-
-    Ok(config)
-}
-
-fn handle_envars(config: &mut Config) -> Result<()> {
-    let re = Regex::new(r"\$([A-Z_]+)").unwrap();
-
-    handle_envar(&mut config.downloads_location, &re)?;
-
-    handle_envar(&mut config.installation_location, &re)?;
-
-    handle_envar(&mut config.version_sync_file_location, &re)?;
-
-    Ok(())
-}
-
-fn handle_envar(item: &mut Option<String>, re: &Regex) -> Result<()> {
-    let value = if let Some(value) = item.as_ref() {
-        value
-    } else {
-        return Ok(());
-    };
-
-    if re.is_match(value) {
-        let extract = re.captures(value).unwrap().get(1).unwrap().as_str();
-        let var =
-            env::var(extract).unwrap_or(format!("Couldn't find {extract} environment variable"));
-
-        *item = Some(value.replace(&format!("${extract}"), &var))
-    }
-
+    cli::start(config).await?;
     Ok(())
 }
